+++ lib/ide/src/controller.rs
@@ -19,6 +19,3 @@ pub mod graph;
 pub mod module;
 pub mod project;
 pub mod text;
-
-/// General-purpose `Result` supporting any `Error`-compatible failures.
-pub type FallibleResult<T> = Result<T,failure::Error>;
