+++ lib/ide/src/lib.rs
@@ -35,6 +35,8 @@ pub mod prelude {
     pub use futures::Stream;
     pub use futures::StreamExt;
     pub use futures::task::LocalSpawnExt;
+
+    pub use utils::fail::FallibleResult;
 }
 
 use crate::prelude::*;
