+++ lib/ide/utils/src/fail.rs
@@ -1,0 +1,4 @@
+//! General-purpose code related to error handling.
+
+/// General-purpose `Result` supporting any `Error`-compatible failures.
+pub type FallibleResult<T> = Result<T,failure::Error>;
