+++ lib/ide/utils/src/lib.rs
@@ -14,4 +14,5 @@ pub use enso_prelude as prelude;
 
 pub mod channel;
 pub mod env;
+pub mod fail;
 pub mod test;
